# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CineAstra API",
    "version": "1.3.0",
    "description": "Fundação e Prompt 02: identidade, sessões revogáveis, TOTP, perfil e privacidade. /api preservado. Bearer JWT; cookies não usados. Banco/migrations e provedores reais são dependências de implantação. Cadastro requer CAPTCHA, termos aprovados e prova de idade; recuperação exige e-mail real. Sem integrações configuradas, 503. Não certifica conformidade legal nem prontidão produtiva. Tarefa14: catálogo administrativo e consultas públicas de filmes/gêneros ativos; uploads/galerias ficam15/16."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Desenvolvimento; runtime usa PUBLIC_ORIGIN validado"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/users/me/children/{id}/authorizations": {
      "post": {
        "operationId": "authorizeMinor",
        "tags": [
          "Identidade"
        ],
        "summary": "Somente responsável com vínculo legal comprovado e ativo",
        "description": "Todos os perfis autenticados; apenas dados próprios e vínculos comprovados, sem bypass ADMIN. Somente responsável com vínculo legal comprovado e ativo. Controle parental deve permitir finalidade; prova de idade vigente. Autorização por operação, prazo de até 24 horas e UNIQUE. Checkout/assinaturas futuros devem consultar requireAge na sua transação. Cache-Control: no-store. IDs em strings. Limites geral e de fluxo podem retornar 429.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              },
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              },
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Caminho do recurso ou acompanhamento conforme descrição da operação"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/IdentityAuthorization"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMedia"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityInput_authorizeMinor"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT HS256 curto com sub/sid/jti, issuer/audience. Sessão persistida, não revogada, usuário ativo e perfil atual. ADMIN/SUPERVISOR e usuários com fator ativo completam 2FA antes de receber acesso."
      }
    },
    "headers": {
      "RequestId": {
        "description": "UUID gerado pelo servidor, nunca eco de header recebido",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "RetryAfter": {
        "description": "Segundos para nova tentativa",
        "schema": {
          "type": "string",
          "pattern": "^[0-9]+$"
        }
      }
    },
    "schemas": {
      "Id": {
        "type": "string",
        "pattern": "^[1-9][0-9]{0,19}$",
        "maxLength": 20,
        "description": "ID INT/BIGINT unsigned decimal string, limite máximo 18446744073709551615; nunca Number",
        "example": "9007199254740993"
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "field",
                "code"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "IdentityInput_authorizeMinor": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "COMPRA",
              "ASSINATURA"
            ]
          },
          "operationReference": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,100}$"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "purpose",
          "operationReference",
          "expiresAt"
        ],
        "additionalProperties": false
      },
...
      "IdentityAuthorization": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "purpose",
          "operationReference",
          "expiresAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "COMPRA",
              "ASSINATURA"
            ]
          },
          "operationReference": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "400: entrada malformada/JSON inválido",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "401: JWT ausente/inválido/expirado ou sessão revogada",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string",
              "example": "Bearer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "403: origem/perfil/dono/vínculo não permitido ou 2FA incompleto",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "404: recurso/rota ausente ou docs desativadas",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "code": "NOT_FOUND",
              "message": "Recurso não encontrado",
              "requestId": "00000000-0000-4000-8000-000000000001"
            }
          }
        }
      },
      "Conflict": {
        "description": "409: conflito de domínio/idempotência",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "413: limite do payload/arquivo",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnsupportedMedia": {
        "description": "415: mídia/encoding não suportado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "422: validação/regra, somente detalhes seguros de campos",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "429: limite geral/IP/identidade/fluxo, registrado em logs",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "500: falha interna, sem SQL/stack/segredo inclusive em development",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "503: dependência indisponível/deadline/shutdown, sem dados internos",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "code": "DEPENDENCY_UNAVAILABLE",
              "message": "Dependência indisponível",
              "requestId": "00000000-0000-4000-8000-000000000001"
            }
          }
        }
      }
    }
  }
}